# [PERF] hr_expense: restrict `manager_id` to internal users