@robodoo 2025-05-14
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/209842.png)](https://mergebot.odoo.com/odoo/odoo/pull/209842)